adds exceptions handling and assets folders
